# app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Comerc",
        "description": "API para teste de emprego na Comerc",
        "contact": {
            "name": "Josué Alessandro"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/clients": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Lista todos os clientes",
                "description": "Obtém a lista de todos os clientes cadastrados",
                "operationId": "4be9bbde54b325142519d9b8ae98b3b4",
                "responses": {
                    "200": {
                        "description": "Lista de clientes"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Clients"
                ],
                "summary": "Cria um novo cliente",
                "description": "Cria um novo cliente com os dados fornecidos",
                "operationId": "65031f51afceced6ba6a7a2b4c0922d7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Client"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente criado com sucesso"
                    },
                    "400": {
                        "description": "Erro na solicitação"
                    },
                    "422": {
                        "description": "Error: Unprocessable Content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/clients/email/": {
            "post": {
                "tags": [
                    "Clients"
                ],
                "summary": "Obtém um cliente por e-mail",
                "description": "Obtém os detalhes de um cliente com base no e-mail",
                "operationId": "88577d9593dbc411f189d3284fdf34a1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "e-mail do cliente",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Detalhes do cliente"
                    },
                    "404": {
                        "description": "Cliente não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/clients/{id}": {
            "put": {
                "tags": [
                    "Clients"
                ],
                "summary": "Atualiza um cliente por ID",
                "description": "Atualiza os detalhes de um cliente com base no ID",
                "operationId": "7d0a5ae9e306858d30e04fc50265d80c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do cliente",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Client"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cliente atualizado com sucesso"
                    },
                    "400": {
                        "description": "Erro na solicitação"
                    },
                    "422": {
                        "description": "Error: Unprocessable Content"
                    },
                    "404": {
                        "description": "Cliente não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Clients"
                ],
                "summary": "Exclui um cliente por ID",
                "description": "Exclui um cliente com base no ID",
                "operationId": "8720346de08f1e46fe28a2a7179d23c1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do cliente",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente excluído com sucesso"
                    },
                    "404": {
                        "description": "Cliente não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Client": {
                "required": [
                    "name",
                    "email",
                    "phone",
                    "birthday",
                    "address1",
                    "postalcode",
                    "city",
                    "state"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "birthday": {
                        "type": "string"
                    },
                    "address1": {
                        "type": "string"
                    },
                    "address2": {
                        "type": "string"
                    },
                    "address3": {
                        "type": "string"
                    },
                    "postalcode": {
                        "type": "string"
                    },
                    "city": {
                        "type": "string"
                    },
                    "state": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}